requirements.txt
python-dotenv
requests
playwright
psycopg[binary]
pytest
pytest-bdd
rich
scrapy